pyproject.toml

# --- Build system configuration

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/unidas.py"]

# --- Project Metadata

[project]
name = "unidas"

version = "0.0.1" # Make sure to bump dascore.__version__ as well!

authors = [
  { name="Derrick Chambers", email="chambers.ja.derrick@gmail.com" },
]

description = "A DAS compatibility library"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering",
]

keywords = ["geophysics", "distributed-acoustic-sensing"]

# --- Dependencies

dependencies = [
    "numpy"
]

[project.optional-dependencies]

extras = [
    "dascore",
    "daspy-toolbox",
    "lightguide",
    "xdas",
    "numpy<2"  # This is due to pyrocko. Remove when it has a new release.
]

test = [
    "pre-commit",
    "pytest",
    "ruff",
    "pooch",
    "pytest-cov",
]

dev = ["unidas[test]", "unidas[extras]"]

# --- URLs for project

[project.urls]
"Bug Tracker" = "https://github.com/unidas-dev/unidas"
"Documentation" = "https://github.com/unidas-dev/unidas"
"Homepage" = "https://github.com/unidas-dev/unidas"

# --- External tool configuration

[tool.ruff]

line-length = 88

# enable certain types of linting
lint.select = [
    "E",
    "F",
    "UP",
    "RUF",
    "I001",
    "D",
    "FA",
    "T",
    "N",
    "NPY",
    "NPY201",
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "__init__.py"
]

# lowest python version supported
target-version = "py310"

lint.fixable = ["ALL"]

# List of codes to ignore
lint.ignore = ["D105", "D107", "D401", "D205", "D200", "D400"]

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

# config for docstring parsing
[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pytest.ini_options]
filterwarnings = [
    # Ignore hdf5 warnings from pytables, See pytables #1035
    'ignore::Warning:tables:'
]

[tool.ruff.format]
# Use `\n` line endings for all files
line-ending = "lf"
